Add builder, clean up code
