--- src/org/typo3/chefci/v2/stages/AbstractStage.groovy
@@ -1,23 +1,0 @@
-package org.typo3.chefci.v2.stages
-
-abstract public class AbstractStage implements Stage {
-
-    def script
-    def stageName
-
-
-    AbstractStage(script, stageName) {
-        this.script = script
-        this.stageName = stageName
-    }
-
-    abstract void run()
-
-    @Override
-    void execute() {
-        script.stage(stageName) {
-            run()
-        }
-    }
-
-}
