+++ src/org/typo3/chefci/v2/stages/HelloWorld.groovy
@@ -1,15 +1,20 @@
 package org.typo3.chefci.v2.stages
 
-public class HelloWorld extends AbstractStage {
+public class HelloWorld implements Stage {
+
+    def script
+    String stageName
 
     HelloWorld(script, String stageName) {
-        super(script, stageName)
+        this.script = script
+        this.stageName = stageName
     }
 
     @Override
-    void run() {
-        // script.echo "Hello World!"
-        printf "hallo welt"
+    void execute() {
+        script.stage(stageName) {
+            script.echo "Hiho!"
+        }
     }
 
 }
