--- src/org/typo3/chefci/v2/stages/HelloWorld2.groovy
@@ -1,20 +1,0 @@
-package org.typo3.chefci.v2.stages
-
-public class HelloWorld2 implements Stage {
-
-    def script
-    String stageName
-
-    HelloWorld2(script, String stageName) {
-        this.script = script
-        this.stageName = stageName
-    }
-
-    @Override
-    void execute() {
-        script.stage(stageName) {
-            script.echo "Hiho!"
-        }
-    }
-
-}
